backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
pydantic
aiosqlite
python-dotenv
apscheduler>=3.10.0
pandas
